consolidated component functions, removed ducks file
